backend_api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosmtplib==3.0.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
black==25.1.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
ConfigArgParse==1.7.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-cache2==0.2.2
fastapi-mail==1.5.0
Flask==3.1.1
flask-cors==6.0.0
Flask-Login==0.6.3
frozenlist==1.7.0
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.37.5
locust-cloud==1.21.8
Mako==1.3.9
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.4.4
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pendulum==3.1.0
platformdirs==4.3.6
propcache==0.3.2
psutil==7.0.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.1
python-socketio==5.13.0
pyzmq==26.4.0
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
zope.event==5.0
zope.interface==7.2
